backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.2
